backend/foodgram/requirements.txt
Django==3.2.15
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
flake8==4.0.1
cryptography==37.0.2
django-colorfield==0.7.2
drf-base64==2.0
gunicorn==20.0.4
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.21.0
fpdf==1.7.2
pytz==2021.3
reportlab==3.6.12
sqlparse==0.4.2